Support taskwarrior 2.6.0 serialization format
As of taskwarrior 2.6.0 the depends field changed from being a string of
comma seperated uuid's to being a proper json array of uuid strings.
This commit allows the selection of which serialization format to use
with the help of typestating. Task<TW26> will (de)serialize to the new format,
while Task<TW25> will retain the older behaviour. This will allow this
library to be used with any version of taskwarrior.

Signed-off-by: Imran Iqbal <imran@imraniqbal.org>